for codex

+++ services/netmanagernative/src/manager/dns_manager.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Huawei Device Co., Ltd.
+ * Copyright (c) 2022 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -57,7 +57,7 @@ int32_t DnsManager::CreateNetworkCache(uint16_t netId)
     return DelayedSingleton<DnsParamCache>::GetInstance()->CreateCacheForNet(netId);
 }
 
-void DnsManager::SetDefaultNetwork(int netId)
+void DnsManager::SetDefaultNetwork(uint16_t netId)
 {
     DelayedSingleton<DnsParamCache>::GetInstance()->SetDefaultNetwork(netId);
 }
